+++ src/hooks/lua_api/mod.rs
@@ -46,6 +46,7 @@ mod serializers;
 pub(crate) mod storage;
 pub(crate) mod template_data;
 pub(crate) mod transaction;
+pub(crate) mod tx_hooks;
 pub(crate) mod utils;
 mod vm_label;